+++ src/SWP/Bundle/BridgeBundle/Resources/config/doctrine-orm/Package.orm.yml
@@ -25,7 +25,7 @@ SWP\Component\Bridge\Model\Package:
         language:
             type: string
         description:
-            type: string
+            type: text
             nullable: true
         subjects:
             type: array
