+++ src/SWP/Bundle/ContentBundle/Hydrator/ArticleHydrator.php
@@ -2,6 +2,18 @@
 
 declare(strict_types=1);
 
+/*
+ * This file is part of the Superdesk Web Publisher Content Bundle.
+ *
+ * Copyright 2016 Sourcefabric z.ú. and contributors.
+ *
+ * For the full copyright and license information, please see the
+ * AUTHORS and LICENSE files distributed with this source code.
+ *
+ * @copyright 2016 Sourcefabric z.ú
+ * @license http://www.superdesk.org/license
+ */
+
 namespace SWP\Bundle\ContentBundle\Hydrator;
 
 use SWP\Bundle\ContentBundle\Model\ArticleInterface;
@@ -9,7 +21,7 @@ use SWP\Bundle\ContentBundle\Provider\RouteProviderInterface;
 use SWP\Component\Bridge\Model\ItemInterface;
 use SWP\Component\Bridge\Model\PackageInterface;
 
-class ArticleHydrator implements ArticleHydratorInterface
+final class ArticleHydrator implements ArticleHydratorInterface
 {
     /**
      * @var RouteProviderInterface
@@ -52,8 +64,6 @@ class ArticleHydrator implements ArticleHydratorInterface
         $article->setMetadata($package->getMetadata());
         // assign default route
         $article->setRoute($this->routeProvider->getRouteForArticle($article));
-        $article->setCreatedAt($package->getFirstCreated());
-        $article->setUpdatedAt($package->getVersionCreated());
 
         return $article;
     }
