+++ src/SWP/Bundle/ContentBundle/spec/Factory/ORM/ArticleFactorySpec.php
@@ -12,28 +12,24 @@
  * @license http://www.superdesk.org/license
  */
 
-namespace spec\SWP\Bundle\ContentBundle\Factory\PHPCR;
+namespace spec\SWP\Bundle\ContentBundle\Factory\ORM;
 
-use Doctrine\Common\Collections\ArrayCollection;
 use PhpSpec\ObjectBehavior;
-use SWP\Bundle\ContentBundle\Doctrine\ODM\PHPCR\Article;
-use SWP\Bundle\ContentBundle\Factory\ORM\ArticleFactory;
 use SWP\Bundle\ContentBundle\Factory\ArticleFactoryInterface;
+use SWP\Bundle\ContentBundle\Factory\ORM\ArticleFactory;
+use SWP\Bundle\ContentBundle\Hydrator\ArticleHydratorInterface;
 use SWP\Bundle\ContentBundle\Model\ArticleInterface;
-use SWP\Bundle\ContentBundle\Model\RouteInterface;
-use SWP\Bundle\ContentBundle\Provider\RouteProviderInterface;
-use SWP\Component\Bridge\Model\Item;
 use SWP\Component\Bridge\Model\PackageInterface;
 use SWP\Component\Storage\Factory\FactoryInterface;
 
 /**
  * @mixin ArticleFactory
  */
-class ArticleFactorySpec extends ObjectBehavior
+final class ArticleFactorySpec extends ObjectBehavior
 {
-    public function let(FactoryInterface $factory, RouteProviderInterface $routeProvider)
+    public function let(FactoryInterface $factory, ArticleHydratorInterface $articleHydrator)
     {
-        $this->beConstructedWith($factory, $routeProvider);
+        $this->beConstructedWith($factory, $articleHydrator);
     }
 
     public function it_is_initializable()
@@ -53,101 +49,16 @@ class ArticleFactorySpec extends ObjectBehavior
         $this->create()->shouldReturn($article);
     }
 
-    public function it_creates_article_from_package_and_fallbacks_slug_to_package_headline(
+    public function it_creates_a_new_article_based_on_package(
         FactoryInterface $factory,
         PackageInterface $package,
-        Article $article,
-        RouteInterface $route,
-        RouteProviderInterface $routeProvider
+        ArticleInterface $article,
+        ArticleHydratorInterface $articleHydrator
     ) {
         $factory->create()->willReturn($article);
 
-        $item = new Item();
-        $item->setBody('some item body');
-        $item->setType('text');
-        $item->setDescription('item lead');
-
-        $package->getHeadline()->shouldBeCalled()->willReturn('item headline');
-        $package->getDescription()->shouldBeCalled()->willReturn('package lead');
-        $package->getBody()->shouldBeCalled()->willReturn('some package body');
-        $package->getItems()->shouldBeCalled()->willReturn(new ArrayCollection([$item]));
-        $package->getLanguage()->shouldBeCalled()->willReturn('en');
-        $package->getMetadata()->shouldBeCalled()->willReturn(['some' => 'meta']);
-        $package->getSlugline()->shouldBeCalled();
-
-        $article->setTitle('item headline')->shouldBeCalled();
-        $article->setBody('some package body some item body')->shouldBeCalled();
-        $article->setLead('package lead item lead')->shouldBeCalled();
-        $article->setLocale('en')->shouldBeCalled();
-        $article->setRoute($route)->shouldBeCalled();
-        $article->setMetadata(['some' => 'meta'])->shouldBeCalled();
-        $article->setSlug('item headline')->shouldNotBeCalled();
-
-        $routeProvider->getRouteForArticle($article)->willReturn($route);
+        $articleHydrator->hydrate($article, $package)->willReturn($article);
 
         $this->createFromPackage($package)->shouldReturn($article);
-    }
-
-    public function it_creates_article_from_package_and_sets_article_slug_from_package_slugline(
-        FactoryInterface $factory,
-        PackageInterface $package,
-        Article $article,
-        RouteInterface $route,
-        RouteProviderInterface $routeProvider
-    ) {
-        $factory->create()->willReturn($article);
-
-        $item = new Item();
-        $item->setBody('some item body');
-        $item->setType('text');
-        $item->setDescription('item lead');
-
-        $package->getHeadline()->shouldBeCalled()->willReturn('item headline');
-        $package->getDescription()->shouldBeCalled()->willReturn('package lead');
-        $package->getBody()->shouldBeCalled()->willReturn('some package body');
-        $package->getItems()->shouldBeCalled()->willReturn(new ArrayCollection([$item]));
-        $package->getLanguage()->shouldBeCalled()->willReturn('en');
-        $package->getMetadata()->shouldBeCalled()->willReturn(['some' => 'meta']);
-        $package->getSlugline()->shouldBeCalled()->willReturn('slugline');
-
-        $article->setTitle('item headline')->shouldBeCalled();
-        $article->setBody('some package body some item body')->shouldBeCalled();
-        $article->setLead('package lead item lead')->shouldBeCalled();
-        $article->setLocale('en')->shouldBeCalled();
-        $article->setRoute($route)->shouldBeCalled();
-        $article->setMetadata(['some' => 'meta'])->shouldBeCalled();
-        $article->setSlug('slugline')->shouldBeCalled();
-
-        $routeProvider->getRouteForArticle($article)->willReturn($route);
-
-        $this->createFromPackage($package)->shouldReturn($article);
-    }
-
-    public function it_throws_an_exception_when_item_type_not_allowed(
-        FactoryInterface $factory,
-        PackageInterface $package,
-        Article $article,
-        RouteInterface $route
-    ) {
-        $factory->create()->willReturn($article);
-
-        $item = new Item();
-        $item->setBody('some item body');
-        $item->setType('fake');
-
-        $package->getHeadline()->shouldNotBeCalled();
-        $package->getBody()->shouldBeCalled()->willReturn('some package body');
-        $package->getItems()->shouldBeCalled()->willReturn(new ArrayCollection([$item]));
-        $package->getLanguage()->shouldNotBeCalled();
-        $package->getMetadata()->shouldNotBeCalled();
-
-        $article->setTitle('item headline')->shouldNotBeCalled();
-        $article->setBody('some package body some item body')->shouldNotBeCalled();
-        $article->setLocale('en')->shouldNotBeCalled();
-        $article->setRoute($route)->shouldNotBeCalled();
-        $article->setMetadata(['some' => 'meta'])->shouldNotBeCalled();
-
-        $this->shouldThrow(\InvalidArgumentException::class)
-            ->duringCreateFromPackage($package);
     }
 }
